% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.em.glm.R
\name{AIC.em.glm}
\alias{AIC.em.glm}
\title{Calcualte the AIC of the em.glm model}
\usage{
\method{AIC}{em.glm}(object, ..., k = 2)
}
\arguments{
\item{object}{A 'em.glm' class returned by the em.glm function.}

\item{...}{optionally more fitted model objects.}

\item{k}{numeric, the \emph{penalty} per parameter to be used; the
    default \code{k = 2} is the classical AIC.}
}
\value{
The AIC score of the model.
}
\description{
Calcualte the AIC of the em.glm model
}
