% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{data.1}
\alias{data.1}
\title{Simulated data set}
\format{An object of class \code{list} of length 3.}
\usage{
data.1
}
\description{
Simulated data set
}
\keyword{datasets}
