% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{deviance.em.glm}
\alias{deviance.em.glm}
\title{Model deviance (calculated from deviance residuals)}
\usage{
\method{deviance}{em.glm}(object, x, y, weight = c(1), ...)
}
\arguments{
\item{object}{An 'em.glm' object.}

\item{x}{An \emph{n}-by-\emph{p} design matrix.}

\item{y}{A vector of observation of length \emph{n}.}

\item{weight}{A \emph{n} length vector of observation weight terms.  This is currently designed to be either the exposure for a Poisson model or the number of trials for a Logistic model.}

\item{...}{other arguments.}
}
\value{
The model deviance statistic.
}
\description{
Model deviance (calculated from deviance residuals)
}
