% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprob.list.R
\docType{data}
\name{dprob.list}
\alias{dprob.list}
\title{List of distribution functions accessed by family name ("poisson" or "binomial").}
\format{An object of class \code{list} of length 2.}
\usage{
dprob.list
}
\description{
List of distribution functions accessed by family name ("poisson" or "binomial").
}
\keyword{datasets}
