% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emax.glm-package.R
\docType{package}
\name{emax.glm}
\alias{emax.glm}
\alias{emax.glm-package}
\title{General linear regression via  Expectation-Maximization.}
\description{
A package for fitting Expectation Maximization algorithms for general linear models.
    The package currently supports Poisson and Logistic regression with variable weights.
    New users are recommended to look at the 'em.glm' and 'small.em' functions.
}
