% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprob.list.R
\name{make.dpois}
\alias{make.dpois}
\title{Build a Poisson log likelihood}
\usage{
make.dpois(x, y, linkinv = poisson()$linkinv, weight = c(1),
  log = FALSE)
}
\arguments{
\item{x}{An \emph{n}-by-\emph{p} design matrix.}

\item{y}{A vetor of observation of length \emph{n}.}

\item{linkinv}{Inverse link function desired}

\item{weight}{A \emph{n} length vector of observation weight terms.  This is currently designed to be either the exposure for a Poisson model or the number of trials for a Logistic model.}

\item{log}{Boolean flag.  If TRUE returns the log dist.}
}
\description{
Build a Poisson log likelihood
}
