% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.em.glm.summary.R
\name{plot.em.glm.summary}
\alias{plot.em.glm.summary}
\title{Error bar plot of coefficients and errors to inspect class overlap.}
\usage{
\method{plot}{em.glm.summary}(x, ...)
}
\arguments{
\item{x}{An EM summary object, from summary.em.glm}

\item{...}{Arguments to be passed to methods}
}
\value{
An R plot with error bars.
}
\description{
Error bar plot of coefficients and errors to inspect class overlap.
}
\examples{
x <- model.matrix(~ factor(wool) + factor(tension), warpbreaks)
y <- warpbreaks$breaks

m <- em.glm(x = x, y = y, K = 2, b.init = "random")
m.sum <- summary(m)

plot(m.sum)

}
