% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.em.glm.summary.R
\name{plot_probabilities.matrix}
\alias{plot_probabilities.matrix}
\title{Plot the class probabilities, both compared to data set index and as histogram.}
\usage{
\method{plot_probabilities}{matrix}(class_probabilities, ...)
}
\arguments{
\item{class_probabilities}{Matrix of n x K class probabilities}

\item{...}{Associated arguments to be passed to plot::par function.}
}
\description{
Plot the class probabilities, both compared to data set index and as histogram.
}
