% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.em.glm.summary.R
\name{plot_probabilities}
\alias{plot_probabilities}
\title{Probability plots for the K classes fit}
\usage{
plot_probabilities(...)
}
\arguments{
\item{...}{Arguments to be passed to methods}
}
\description{
Probability plots for the K classes fit
}
