% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{results_k25_n1000}
\alias{results_k25_n1000}
\title{Simulated data set}
\format{An object of class \code{matrix} with 3 rows and 50 columns.}
\usage{
results_k25_n1000
}
\description{
Simulated data set
}
\keyword{datasets}
