% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{results_simple}
\alias{results_simple}
\title{Simualted data set}
\format{An object of class \code{matrix} with 3 rows and 50 columns.}
\usage{
results_simple
}
\description{
Simualted data set
}
\keyword{datasets}
