% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{sim.1}
\alias{sim.1}
\title{Simulated data set}
\format{An object of class \code{list} of length 2.}
\usage{
sim.1
}
\description{
Simulated data set
}
\keyword{datasets}
