% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{sim.2}
\alias{sim.2}
\title{Simulated data set}
\format{An object of class \code{list} of length 3.}
\usage{
sim.2
}
\description{
Simulated data set
}
\keyword{datasets}
