% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{sim.3}
\alias{sim.3}
\title{Simulated data set}
\format{An object of class \code{list} of length 5.}
\usage{
sim.3
}
\description{
Simulated data set
}
\keyword{datasets}
