% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_probabilities.R
\name{update_probabilities}
\alias{update_probabilities}
\title{Construct normalized class properties for a given set of parameters}
\usage{
update_probabilities(dprob, params)
}
\arguments{
\item{dprob}{Probabiltiy distribution function to call.  See 'dprob.list' for examples.}

\item{params}{List of class parameters.  Length of list is number of classes}
}
\description{
Construct normalized class properties for a given set of parameters
}
