% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.em.glm.R
\name{IC.em.glm}
\alias{IC.em.glm}
\title{General Information Criteria function}
\usage{
IC.em.glm(em.glm, alpha)
}
\arguments{
\item{em.glm}{An emax glm fit.}

\item{alpha}{Scaling factor for information criteria (2 or ln(\emph{n}) for AIC and BIC respectively).}
}
\value{
The IC value of the model for given value of k.
}
\description{
General Information Criteria function
}
