% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.em.glm.R
\name{IC.em.glm}
\alias{IC.em.glm}
\title{General Inforamtion Criteria function}
\usage{
IC.em.glm(em.glm, alpha)
}
\arguments{
\item{em.glm}{An emax glm fit.}

\item{alpha}{Scaling factor for inforamtion criteria (2 or ln(\emph{n}) for AIC and BIC respectively).}
}
\description{
General Inforamtion Criteria function
}
