% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.em.glm.summary.R
\name{plot_probabilities.em.glm}
\alias{plot_probabilities.em.glm}
\title{Test Plot em.glm}
\usage{
\method{plot_probabilities}{em.glm}(em.glm, ...)
}
\arguments{
\item{em.glm}{An em.glm object.  From em.fit}

\item{...}{Associated arguments to be passed to plot::par function.}
}
\description{
Test Plot em.glm
}
\examples{
x <- model.matrix(~ factor(wool) + factor(tension), warpbreaks)
y <- warpbreaks$breaks

m <- em.glm(x = x, y = y, K = 2, b.init = "random")

plot_probabilities(m)

}
